Fixed printing of times, change ComputationTimeList in Float

+++ src/ExpressionResponse/ResultResponse.java
@@ -2,7 +2,7 @@ package ExpressionResponse;
 
 public class ResultResponse extends Response{
 
-    public static String create(String result, long time) {
+    public static String create(String result, float time) {
         return "OK;" + time + ";" + result;
     }
 }
